fix(BaseService): restore POST functionality w.r.t attack_id
The base create method now permits users to include workspace.attack_id
as long as (1) we're creating a new snapshot as opposed to a new object;
(2) server-controlled fields specified in the req body match the prev
snapshot. This will restore support for the POST endpoint tangential
use-case of creating new object snapshots as opposed to exclusively
creating new objects.